+++ app/src/main/java/com/example/todolist/feature/todolist/domain/use_case/task_item/GetTaskItemById.kt
@@ -8,7 +8,7 @@ class GetTaskItemById(
 ) {
 
     @Throws(InvalidTaskItemException::class)
-    suspend operator fun invoke(id: Int) {
+    suspend operator fun invoke(id: Long) {
         repository.getTaskItemById(id)
     }
 }
